+++ radio_interferometer/__init__.py
@@ -1,3 +1,3 @@
 """Radio interferometry FX correlator application."""
 
-__version__ = "0.1.8-dev"
+__version__ = "0.2.0-beta"